requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-model-utils==4.1.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
